requirements.txt
requests==2.26.0
python-dotenv==0.21.1
pyppeteer==1.0.2
beautifulsoup4==4.12.2